The Opengl renderer now renders colors for each vertex correctly.  The DX9 and OpenGL renderers should now be in sync.

+++ code/core/time/Time.cpp
@@ -23,7 +23,7 @@ Time *Time::GetInstance()
 
 void Time::Start()
 {
-	mfStartTime = timeGetTime();
+	mfStartTime = static_cast<float>(timeGetTime());
 }
 
 void Time::End()
